AgroGestao-main/requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.11.0
iniconfig==2.3.0
mysqlclient==2.2.8
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
sqlparse==0.5.5
tzdata==2025.3
